requirements.txt
Flask>=2.0
requests>=2.20
google-api-python-client>=2.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=0.19